info file added
Wasn't exactly sure how to format it since the illustrations have different titles. Either way, it should be clear to anybody who reads it.
Also wasn't sure what other information I could add-on, but I would like my twitter in there somewhere if possible. (and if y'all are ok with it)